refactor(test): migrate test_chained_client unique tests to test_ai_chained_client
迁移 2 个唯一测试（chain 解析和未知 provider 校验）至 test_ai_chained_client.py，删除旧文件
